requirements.txt
deepface
opencv-python
numpy
tf-keras
fastapi
uvicorn
requests
pydantic
python-dotenv
